Trigger backup from main database process.
Sqlite backups from the same database connection are much more efficient!
Externally triggered backups restart every time the database changes during
a backup, or need to lock the database exclusively for a long time.

We periodically look for a "backup.request" file, and if it exists
will write a backup. We write to "backup.tmp" first, and when finished
rename the file to "backup.db".

A crontab job can touch the "backup.request" file and then wait for the
"backup.db" file to appear.